requirements.txt
requests==2.31.0
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.6
faiss-cpu==1.7.4
sentence-transformers==2.2.2
python-dotenv==1.0.0
google-generativeai==0.3.2
